feature: parse multilingual px files and output metadata as json
The package got refactored. It is now able to digest also multilingual
px files. The dataframe and the metadata will be localized to any
given locale that is inluded in the px cubes `LANGUAGES` parameter.

When an output directory is specified, the output can also be written to a files.
The following files will be written in that case:

- metadata.json: px records redunance free as json file
- <language>.json for all languages the px file was translated into
- metadata-<locale>.json if a locale is given
- data-<locale>.csv with the data and headers in the given locale

The package still supports only ANSI encoded px cubes. Also a list of supported
keywords is inlcuded with the package: `supported_keywords.csv`. In case an unsupported keyword is
encountered this will be reported.

The file to parse can be either an url or a file that was previously downloaded.

The package was refactored in order to provide data and metadata in a redundance free way:
The workflow now is as follows:

1. The files is scanned and read as lines separated by `;` in the px cube file
2. The lines are then grouped into a list of px rows, that always consist of a px keyword and the
content for that keyword: more line can belong to the same px row group
3. All px rows that are not translations are then processed to derive at the metadata in the default
language of the cube
4. Next all px rows are used to extract string translations for all strings in the px cube files, that come
with translations: strings such as years and codes that are the same in all languages are not included in
the extracted translations.
5. The translations and the default metadata are then used to localize the metadata
6. The localized dataframe is derived from the data and localized metadata.

The internal structure of the metadata and data matches what is written to the output files.
In the `tests/testthat/` example data and output that is added for testing purposes.