feature: add parsing for multilingual px cube files
Add parsing for multilingual px files.
The dataframe and the metadata will be localized to any
given locale that is inluded in the px cubes `LANGUAGES` parameter.

When an output directory is specified, the output can also be written to a files.
The following files will be written in that case:

- metadata.json: px records redunance free metadata as json file
- metadata-<locale>.json if a locale is given
- data-<locale>.csv with the data and headers in the given locale

The package still supports only ANSI encoded px cubes. Also a list of supported
keywords is included with the package: `supported_keywords.csv`. In case unsupported
keywords are encountered: these will be reported.

The file to parse can be either an url or a file that was previously downloaded.

The parsing workflow is as follows:

1. The files is scanned and read as lines separated by `;` in the px cube file
2. The lines are then grouped into a list of px rows, that always consist
of a px keyword and the content for that keyword
3. All px keys for the default language are selected
4. These rows that are then processed to derive at the default metadata:
(the metadata in the default language)
5. Next all px keys for a locale are selected
6. These keys are then processed to arrive at the localized version of the
metadata
7. The localized dataframe is derived from the data and localized metadata.

The internal structure of px files and the output files can be seen
at `tests/testthat/` where complete input and output files have been
added for testing purposes.