refactor: rename some functions and scripts and fix bugs
functions are renamed and regrouped into script, so that the
main functiuons are in their own scripts and the grouping
of the helper functions by topic is made more consistent.

Also bugs regarding the translations for dimension dependent
keys are fixed: example `PRECISION("Indikator","Scheidungen")
has the English translations stored at the key:
`PRECISION[en]("Indicator","Divorces") where a flag `[en]`
indicates a translation and the dimension `Indikator`
and the value `Scheidungen` both appear in their translated
form.
Recognizing these edge cases for how to find the translations
in the px file has been improved.