bug fix: encoding
it seems safer to use just the given input parameter `encoding`
for both `readLines` and `scan`.

`scan` also offeres a parameter `fileEncoding` that overwrites its
parameter `encoding`. Since `CODEPAGE` includes a file encoding in
the px files with a default, it seemed a good solution to use
the fileEncoding for `scan`. But surprisingly this led to decoding
bugs in some cases, therefore this commit is going back to just
using the `encoding` from the inpur parameter in both cases: `scan`
and `readLines`