refactor the package: code changes
the goal is to make the parsing faster and more reliable
the refactoring wants to make the package a good addition the BFS
package for py data: it can take care of cases of large py files
and delivers the data along with the most important metadata in a
fast and reliable way.

The feature to write the output on file was removed. Also only the
dimensions and some minimal metadata are collected, as the bugs
in the previous version of the package were related to some edge
cases of metadata collection on metadata that was of questionable
interest